+++ OpenSim/Tools/Configger/ConfigurationLoader.cs
@@ -70,7 +70,7 @@ namespace OpenSim.Tools.Configger
 
             List<string> sources = new List<string>();
 
-            string masterFileName = startupConfig.GetString("inimaster", "OpenSimDefaults.ini");
+            string masterFileName = startupConfig.GetString("inimaster", "OpenSim.ini");
 
             if (masterFileName == "none")
                 masterFileName = String.Empty;